Corregir validaciones en comentarios

+++ Love4AnimalsApi/Services/CommentService.cs
@@ -5,36 +5,54 @@ namespace Love4AnimalsApi.Services;
 public class CommentService : ICommentService
 {
     private ICommentRepository commentRepository;
-    public CommentService(ICommentRepository commentRepository)
+    private IPostRepository postRepository;
+    private IUserRepository userRepository;
+    public CommentService(ICommentRepository commentRepository, IPostRepository postRepository, IUserRepository userRepository)
     {
         this.commentRepository = commentRepository;
+        this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
-    public List<GetCommentDto> GetComments(long postId)
+    public List<GetCommentDto>? GetComments(long postId)
     {
+        Post? post = postRepository.GetPost(postId);
+        if (post == null) return null;
         List<Comment> comments = commentRepository.GetComments(postId);
         return comments.Select(c => new GetCommentDto(c.Id, c.Content, c.CreatedAt, c.UserId, c.PostId)).ToList();
     }
     public GetCommentDto? GetComment(long postId, long id)
     {
+        Post? post = postRepository.GetPost(postId);
+        if (post == null) return null;
         Comment? comment = commentRepository.GetComment(postId, id);
         if (comment == null) return null;
         return new GetCommentDto(comment.Id, comment.Content, comment.CreatedAt, comment.UserId, comment.PostId);
     }
-    public GetCommentDto CreateComment(long postId, CreateCommentDto createCommentDto)
+    public GetCommentDto? CreateComment(long postId, CreateCommentDto createCommentDto)
     {
+        Post? post = postRepository.GetPost(postId);
+        if (post == null) return null;
+        User? user = userRepository.GetUser(createCommentDto.UserId);
+        if (user == null) return null;
         Comment comment = new(0, createCommentDto.Content, DateTime.Now, createCommentDto.UserId, postId);
         Comment createdComment = commentRepository.CreateComment(comment);
         return new GetCommentDto(createdComment.Id, createdComment.Content, createdComment.CreatedAt, createdComment.UserId, createdComment.PostId);
     }
-    public GetCommentDto? UpdateComment(long id, UpdateCommentDto updateCommentDto)
+    public GetCommentDto? UpdateComment(long postId, long id, UpdateCommentDto updateCommentDto)
     {
-        Comment comment = new(id, updateCommentDto.Content, DateTime.Now, 0, 0);
+        Post? post = postRepository.GetPost(postId);
+        if (post == null) return null;
+        Comment comment = new(id, updateCommentDto.Content, DateTime.Now, 0, postId);
         Comment? updatedComment = commentRepository.UpdateComment(id, comment);
         if (updatedComment == null) return null;
         return new GetCommentDto(updatedComment.Id, updatedComment.Content, updatedComment.CreatedAt, updatedComment.UserId, updatedComment.PostId);
     }
-    public bool DeleteComment(long id)
+    public bool DeleteComment(long postId, long id)
     {
+        Post? post = postRepository.GetPost(postId);
+        if (post == null) return false;
+        Comment? comment = commentRepository.GetComment(postId, id);
+        if (comment == null) return false;
         return commentRepository.DeleteComment(id);
     }
 }
